requirements.txt
django-mathfilters
django-simple-captcha
djangorestframework-simplejwt
django-cors-headers
python-dotenv